Remove unused .csslintignore
This file was for the Jenkins job of csslint (since it didn't
use the Gruntfile), now that it does, this file is redundant.

Note that .jshintignore should not be removed because, contrary
to csslint, jshint actually supports the ignore file natively,
so .jshintignore is used by Grunt as well (wheareas in our
csslint setup we had to duplicate it).

Change-Id: I6fdb22a382d8aba09b5a73b92866288d2ae70c7b