Fixes for tool group re-population
When registering a tool after a tool group was built, re-population
would end up removing all tools in the group - this was because we
weren't considering tools already assigned to the group to be
"available".

Also, when requesting a tool by name which is not available, the system
would explode, we now check that the tool exists before proceeding with
telling the factory to create it.

The original patch (I28905b195d5a616363780cccbdfd3aa1cc463f56) didn't
really solve the problem completely, but did motivate me to sit down and
fix this, so that's positive.

Thank you to Eranroz.

Change-Id: I937a0cc42ff141db594a804e6577cfb2dcf1a6e5