Moved PNG icons to their own variant
As a humorous oversight, the way variants work in the oojs-ui grunt
build script is such that default is the base, and variants are added
onto default. This means that while the default had PNG icons, the
SVG variant had BOTH!

By splitting the icons out into PNG and SVG variants, we can ensure
that we aren't sending both. I suspect there's a better way to do this,
such as defining the resources differently so we know when something is
mutually exclusive.

Change-Id: Id4a7d3641cfab7a1c2ec7f2ddf4a6e56afb5b271