Removed some of the drop-shadows on buttons
There were a lot of rendering inconsistencies between browsers, and it
wasn't clear that the shadows were of any value, so we just remove them
and the problem goes away.

And that's the way we achieve our goals in VisualEditor - by redefining
the meaning of success.

Also, made text inputs white instead of gray when not focused - which
feels like the right thing to do.

And that how we decide to change things in VisualEditor, with extensive
research and hard evidence.

Change-Id: I19d33d395c9d8db59fba6fa03d87a642c9cc7c91