build: Concatenation task should not include 'default' in a variant
In general it wouldn't include the default file for a variant-based
filepath. But as it does lazy discovery of new variants (at which
point it creates a new buffer based on the defaults' buffer so far),
this sometimes happens if it iterated over the 'default' property
of the current file path variant first.

So it would only at most include the default variant in addition
to the correct variant of the last-iterated file, it would never
duplicate include all files from default and the variant, only
the first.

Fixed by saving the buffer before the individual file variant
iteration starts.

This reverts commit 27832b1dd262bb17f21e2bb747482b153101c4a9.

Change-Id: I5b9ae4a8b1a42ac7e6f4142aff2d48daf782cc1c