[BREAKING CHANGE] Split styles up into base styles and a theme
The impetus of this work was to make way for alternative themes,
especially ones that are designed for different devices and skins.

The theme is called Apex, because it's been designed together with the
MediaWiki Apex skin that's floating around in a sandbox repository
somewhere. The theme works well with the MediaWiki Vector skin too.

Straightforward split with a couple of exceptions:
* ToggleSwitchWidget's glow element is hidden/shown in core but faded in
  and out in apex
* ButtonedElement's hiding of icons and indicators is done inside the
  individual rules rather than introducing a group one, smaller output and
  simpler code

This change moves the theme styles to a separate CSS file called
oo-ui-apex.css which must now be included by VisualEditor.

Using no theme is not recommended.

Change-Id: I2441cf0e869d62b81a722941ba6530de7642d13b