Add maxrows option to OO.ui.TextInputWidget
Once upon a time (2009) in a land far far away (Wikia satellite office)
there was an engineer (me) who for the life of him couldn't figure out how
to get the browser to tell him how high a line of text in a text area was.
Engineers before him struggled with this very same problem, and eventually
just hard-coded the size in pixels, and even special-cased that
measurement for different operating systems. Even as engineers from all
around pitched in, for the life of them, the problem went unsolved.

Many years later (2014) while solving a different problem, the engineer
(me) was again faced with the same problem. How high is a row of text in a
textarea? Then, suddenly, he had an epiphany. He not only came up with a
solution, but instantly realized that the solution to his new and old
problem were one and the same.

For it was the humble textarea element and its oft overlooked rows
attribute which would eventually prove invaluable. By simply setting the
rows attribute to different sizes and then measuring the differences, one
could easily apprehend the exact size of a row of text in a textarea. A
solution both elegant, and unfortunately heretofore evasive.

The engineer (me) took comfort in knowing that at least he eventually
found a solution. He wrote a patch to control the maximum size of a
textarea widget in rows, pushed it into Gerrit, and lived happily ever
after.

The end.

Change-Id: I548446ea3eb5d68907405378acc1a9f02ae8af22