Don't use transitions for opacity on hover for now
There's a bug in Chrome where a lot of content around the transitioned
element will "wiggle" (move a few pixels, then move back again).
The effect we're trying to achieve is so minor that we can afford
to just not use a transition here temporarily while we wait for the
browser bug to be fixed (if that ever happens).

Bug: 62591
Change-Id: Ied6debd7fae5a545736492776a1fd4a9e85f1f62