apex/button: Reset border-color of disabled button when not hovered
We already reset the border-color a few lines down in the :hover
rule, but due to the coloring for the constructive flag etc. now
being in this theme file (and before the disabled rule) that color
survives into the non-hover state as well.

I tried moving it from :hover to the main rule, but that meant
the darker green (from constructive:hover) got exposed on hover
instead creating a different bug. It needs to be in both in order
to have the right (higher) cascading specificity.

Also add a demo entry for a button that is both flagged and
in the disabled state.

Bug: 60498
Change-Id: If08bcadb5da39a70bbc56cc98ec742b03ea89bb1