Only blur inputs within booklet pages when moving away
When moving from a page with an input to a page without one, focusing
the input of the next page is not possible, thus the old page's input
remains focused. To fix this we blurred any focused input on the
previous page - but when setPage is used to set the already current
page, we end up blurring unnecessarily and then quickly focus again, and
that causes problems in Firefox, because Firefox sucks.

Bug: 62841
Change-Id: Ic9e571189ca5257ee0a31db84e45e1ce092b9b8f