Followup 099f26ec: fix $content in FieldsetLayout
When you create a FieldsetLayout with the '$content' config option,
you expect that content to go below the label, not above it. However,
the constructor uses .append() after invoking the parent constuctor,
so the label and icon get inserted at the end instead.

This used not to be a problem because the icon is absolutely positioned
(so its DOM position doesn't matter) and the label was a <legend>
(so it's always treated as if it's the first thing), but when 099f26ec
made the label a <div>, this weird DOM order started to break things.

Bug: 64013
Change-Id: Ib54505cc71d258fed2a537b7cef8d076a215dbe1