PanelLayout: Remove call to getDir()
Was introduced in I418e669b09995521 but doesn't appear to be needed
any more.

As used in the ConfirmationDialog, this gets called when the
iframe is contained in a "display: none;" <div> of OO.ui.Window.

Retrieving style properties after the point in Frame where we
remove "visibility: hidden;" and add "display: none;" is both
pointless (as the calculation will not be accurate) and
problematic as it can cause Firefox to throw an exception due to
a bug in jQuery (without that bug it would return null, which is
just as useless)

Bug: 65373
Change-Id: I31b495fc2d9725f2629a2b53145e12abe6ae1311