Allow FieldLayouts to have help text
For addition of help text in VE page settings dialog

Text shows as a popup when you click a '?' icon next to the field

Mainly uploading now to check that this is actually what we want to do.
Need to find a way to make the popup box fit the text nicely.

Bug: 51798
Change-Id: I30b9a5ce0c73d181910ee832ddc8a98c7d516890