cleanup: Use local var instead of bind() for inline functions
Using .bind() makes sense for binding a referenced callback, but
when the function is inline, might as well use a local variable.

* Reduces code complexity (less boilerplate and wrappings around
  the callback soup).
* Makes code slightly faster (less function call overhead ahead
  of time).
* Easier to read (only one 'this' per scope, no redefinitions of
  'this' to keep track of).
* Easier to maintain (it's hard to detect when a bind is done for
  no reason, e.g. an unused 'this'. An unused variable 'x = this'
  is picked up by JSHint on the other hand).

In terms of memory holding, the result should be the same as the
inline function naturally holds on to the containing closure
anyway due to lexical scope.

Replaced all instances of "bind( function () { .. } )" in src/.
Left instances of "bind( identifier )" alone.

Change-Id: I6890d959edb3d1fde8d95173d4d891560a61b1c6