Support nested style transplantation
When we're transplanting <style> nodes that were generated
by transplantation, reuse the pollNodeId they already have.
When we're transplanting non-transplanted nodes from a document
that has transplanted nodes in it, we need to account for
them by starting with a higher pollNodeId.

Keep track of these things by adding private properties to
the document and the style nodes generated by transplantStyles,
and check them in future runs to detect nesting.

Change-Id: Ib9f2b6c055e0a101f5f1971112cb99768ab3acd7