[BREAKING CHANGE] Set active and pressed on the buttoned element, not the button
This inconsistency between this and framed/frameless made the styles
quite confusing. This will give us more control over the styling,
allowing changes to the button's outer frame to be modified during
pressed and active states as well as the inner button.

While technically a breaking change, no code in any repository will
actually be affected by it.

Change-Id: Idd8b7d032409ce7d65e7536b5ec6e76379a0b8c8