build: Switch from grunt-recess to grunt-contrib-less
This moves us from less version 1.3 to 1.7, desired to use some of the newer
features of less for forthcoming changes, and away from a build / development
dependecy on an unmaintained tool (last release over a year ago).

Using the 'ieCompat' build option makes no difference currently, but should avoid
accidental mistakes in achieving MSIE8 compatibility. The outputted CSS files are
slighly less pretty in the synchronised formatting of browser-specific attributes
for alignment purposes, and the order of attributes within the selectors has been
normalised, but otherwise appears to be a no-op change.

Change-Id: I417c8c10abca58424012861c66395ea219e8f9ab