Fix label text in RTL demo listing containing parentheses
http://i.imgur.com/LtSf9ga.png

Parentheses are "neutral" and as the last paren. is not considered
belonging to the English text, it is assumed to be right-to-left
(like the surrounding paragraph).

Since it doesn't look like the demos will ever be localized, it is
safe to tell the browser that the whole label is left-to-right text.

Change-Id: I2f0a6190921c5e4b556b463b92b32fd3b36eebcb