Reorganize theme system
Previously we built a stylesheet for core styles and another for each
theme. We also made each theme fully mirror the file structure of the
styles folder, and maintained exhaustive lists of stylesheets in JSON.

We now use LESS @import statements to include all the files needed to
build a theme, which composes base styles and theme specifics into a
unified stylesheet.

Here's how it works:

* The base theme specifies variables, mixins, rules and a set of empty
  mixins which are invoked at the end of each base ruleset
  (styles/common.less)
* The base theme's core file imports all of its individual files
  (styles/core.less)
* A theme specifies variables, mixins and theme overrides for the empty
  mixins specified by the base
  (themes/[theme name]/common.less)
* A theme includes base and local common files, includes its theme mixin
  overrides, then includes the core rulesets which use the theme mixin
  overrides
  (themes/[theme name]/core.less)

This change also:
* Renames agora to minerva, since it was designed to be used with the
  minerva skin, and does not conform to agora designs
* Combines what were mostly very small files into logical groups, making
  theme development much less labor-intensive
* Introduces a blank theme template for authors of new themes to start
  with

Change-Id: Ia357bb38891f57f2f8f724f4e95cceada462a4b5