Fix click-block focus-loss prevention
We wanted to avoid clicking on the click block from transferring focus
from the iframe of an isolated window to the top-level frame where the
window is attached. To accomplish this, we simply canceled all mouse
down events on the window manager. Because the mousedown events within
the iframe did not bubble and sink through the iframe barrier, this
caused no side effects within the actual window content.

Now that we have non-isolated windows, this is a problem. The result is
that all mousedown events that aren't explicitly handled at a local
level are canceled - so buttons would work, but text selection or
clicking into an input would not.

This patch moves the event canceling code to the window, so it can
specifically target the click blocking div, and also filters the events
to ensure that only mousedown events on the actual click blocking div,
and not just one of its children, are canceled.

Change-Id: Id45ea41affb2619a719ebba079f1379eb8c6939f