Make CSS transplantation work in IE10
When ResourceLoader creates synthetic <style> nodes,
it does that using code like this in IE:

style = document.createElement( 'style' );
head.appendChild( style );
style.styleSheet.cssText = text;

The usual strategy of appending a text node to the <style>
node doesn't appear work in IE. However, setting .cssText this
way means that a TextNode is not created in the DOM
in IE10 (but it is in IE11).

So, as a fallback, look at styleSheets[i].cssText if
styleNode.textContent is empty.

Change-Id: I6d3bfae6630b639933eaf5e7ac438e35e04e9548