Followup 67f83f1: fall back to empty string to avoid 'undefined'
Apparently we sometimes have <style></style> in the DOM. These
nodes have .textContent = '' , which is falsy so it falls back
to .cssText, which is undefined in non-IE browsers, so we
produce CSS like 'undefined #foo { font-family: bar; }" which
is invalid, so that CSS never loads, so windows never finish
loading and time out. Sigh.

Bug: 70467
Change-Id: I7bcd853ac852d41acaf7c198a1b9515e102e4da4