Work around position: fixed; bug in Safari
Apparently Safari doesn't like it when you nest two
elements that are both position: fixed;. Mouse clicks
would be mapped to the wrong place if the viewport
was scrolled down.

Work around this by keeping dialogs as position: fixed;,
but making the window frame inside of the dialog
position: absolute; instead.

Bug: 70578
Change-Id: I4a2ae78ab351bd313431d8caff595921c8f54a1f