Fix automatic sizing of windows
We were using scrollHeight to measure the ideal height
of the window's body, but if the window is already too tall,
that doesn't work because scrollHeight returns
max(idealHeight, actualHeight).

Work around this by temporarily resizing the window's
frame to height: 1px; for the getBodyHeight() call.

Bonus: don't call updateWindowSize() twice. There's no
point calling it after ready in addition to after setup,
because visible DOM changes in the ready process
(especially ones that affect the height) are evil: the
window is already visible at this point so doing this
would cause a visible flash.

Bug: 70016
Change-Id: I6641f5babb8fd4a426bb4b1d1e292c77801d5ad4