PopupWidget: Make $element not be a 0x0 box
Needed for I4e49a466 in VisualEditor.

We need to make the dialog body align correctly under the anchor.
However, $element is intended to be positioned where popup's anchor
points, and therefore we can't position it from CSS or JS.

Previously the both the anchor and the body were positioned absolutely,
which naturally aligned the anchor with the top, but gave the popup
with no flow content.

Now the body is positioned relatively, making it flow content for the
popup. The anchor is still positioned absolutely, with a 'top' rule
necessary to account for the body not being positioned absolutely
anymore.

Bug: 52526
Change-Id: Ib07a836241ec02aecb0449995f59e0f0e936ccba