Make this.$.$iframe actually point to an iframe again
Window sets its this.$.$iframe to this.$element, but with
the recent-ish refactor that's now a div, not an iframe.
Set it to the actual iframe instead.

This broke the logic in TextInputMenuWidget that compensates
for the menu and its text input being in different frames.
This kept working by accident for non-modal windows because
their $element and $iframe have the same coordinates.

Change-Id: I46a64001152c51b41cd13c2b4100785d4ceb24a8