ClippableElement: Use 'overflow-[xy]: scroll' rather than 'auto'
There is a curious wide-spread browser behavior [1] (can't tell if bug
or feature) where applying explicit height and 'overflow-y: auto' to
an element with no explicit width set doesn't cause the width to be
recalculated, inserts the vertical scrollbar, and thus causes a
horizontal scrollbar to also be necessary.

This doesn't happen with 'overflow-y: scroll' (which causes the width
to be enlarged to fit the scrollbar), and we can safely use it here as
we already know that the scrollbar will be necessary.

[1] Affects all browsers I tested except IE 11.

Bug: 65059
Change-Id: I330fd8c923d99af286cb10a1df428a62fdfbf3ae