Element: Remove test for deprecated onDOMEvent()
The onDOMEvent() method is a transparent proxy for jQuery#on.
It doesn't provide any logic that is in need of testing.

Due to the test being very specific, it actually fails on Firefox.
(On Firefox the iframe receives a focusin event even after we
detach it from the document, so the assertion fails.)

Change-Id: I89920e7d6da234ec1af2cdbb86558c640ebcae4b