Position popups and their anchors correctly in RTL
For both the popups themselves, and the anchors within, we
were relying on position: absolute; to put them at the top
left of their offset containers. This works in LTR, but RTL
it puts them at the top right. Add left: 0; to ensure they
are at the top left, which is needed to make offset computations
work as expected.

Also noflip the margin-left in PopupTool, which is being
used for alignment in a similar way.

Change-Id: I9ae783576699836c6c209179e4474356090abc7b