Conf for doxygen based PHP doc
MediaWiki core is using Doxygen to generates it is documentation. It is
reasonably fast.

Provide a Doxyfile generated with version 1.8.8 with all comments and
default settings stripped out.

Assumes Graphviz is installed by setting HAVE_DOT = YES.

The documentation is generated by running doxygen at the root of this
repository, the doc is made available at docs/php/index.html

Bug: 72454
Change-Id: I3377a370bad65307968dd3f90433277daffebabb