[BREAKING CHANGE] Apply font-size: 0.8em; to non-isolated windows too
The Apex and MediaWiki themes applied it to isolated windows and overlays,
but not non-isolated windows. This is difficult for integrations
to work around, because applying font-size: 0.8em to non-isolated
window managers will cause the overlays' font size to be 0.64em.
Instead, just remove this ridiculous exception.

Move the demo's windowManager outside of .oo-ui-demo-container:
it being inside was the hacky way that the demo was using to
apply 0.8em to non-isolated windows. Those windows' overlays
had 0.8em applied twice for 0.64em, but we didn't notice
because none of our demo windows contain complex widgets
that use overlays.

Change-Id: Id25907c4c41b72c879e4010208ed66a3580374ad