Use README.md as doxygen main page
From doxygen documentation:
http://www.stack.nl/~dimitri/doxygen/manual/config.html#cfg_use_mdfile_as_mainpage

USE_MDFILE_AS_MAINPAGE

If the USE_MDFILE_AS_MAINPAGE tag refers to the name of a markdown file
that is part of the input, its contents will be placed on the main page
(index.html). This can be useful if you have a project on for instance
GitHub and want to reuse the introduction page also for the doxygen
output.

Adjusted the markdown to use fenced code block. Confirmed Github
recognize it properly via a gist.

Change-Id: I812a32f0bd11f05aa10a73db334bccd49875adca