MessageDialog: Fit actions again when the dialog is resized
We wait for resize transition to end before doing it. This means that,
in case where the resize causes the labels to actually be re-fit, we
will have two separate 'resize' events and two separate resize
transitions (first width and maybe height, then height).

This is unfortunate, but in practice will probably never be visible
unless we manually resize message dialogs [1], so let's not do that.
It could be fixed properly, but that would require too much tinkering
and a possible performance degradation in the common case.

[1] The only way a user can cause a dialog to resize is by resizing
    browser window, which only resizes dialogs to/from 'full' size,
    which means re-fitting labels will not change the dialog height,
    since it's already 100%.

Bug: T75783
Change-Id: Id76e41b4eeb5689c367928dd320ec65c30bd9e2e