[BREAKING CHANGE] CheckboxInputWidget: Allow setting HTML 'value' attribute
Previously we were using the #setValue/#getValue methods (and 'value'
config parameter) to set checkedness state of the input, rather than
its HTML 'value' attribute. This was just fine in JS, but we need to
be able to set the value in server-side usage in PHP.

Like RadioInputWidget, CheckboxInputWidget now has two new methods:
#setSelected and #isSelected. These need to be used where #setValue
and #getValue were used previously.

Bug: T76645
Change-Id: I32608c7f3f3c20ab5c0c6308d62458175a171b3b