[BREAKING CHANGE] FieldLayout: Handle 'inline' alignment better
If the label was longer than the available space, it would move to the
next line of text due to 'display: inline-block'. But we can't remove
that, or the wrapped label would wrap below the checkbox.

Use some 'display: table' magic to make this work like we want it to.
This forces us to insert another wrapper <div/> in FieldLayout code.

This change is breaking for code that used CSS selectors like
'.oo-ui-fieldLayout > .oo-ui-labelElement-label', since the label and
field are no longer direct descendants of the element.

Took this opportunity to do some CSS cleanup.

Bug: T78030
Bug: T78173
Change-Id: Ie27bd94c57bba763eabdcde431136bd49a4dbcc4