[BREAKING CHANGE] Split primary flag into primary and progressive
Using the primary was specifying both the blue styling to indicate
its function and the primary placement in dialogs to indicate its
importance, but in fact these are separate things that should have
separate names.

Now "primary" is now split, so that "progressive" is used for the
blue styling and "primary" only hints at the importance. In the
case of the MediaWiki theme, buttons that are progressive,
constructive or destructive are only given the strong treatment of
a colored background if they are also primary.

Bug: T78040
Change-Id: Iee91d2c695f7fef3f78a2d44f05b431dbe30615d