Temporarily remove position:absolute on body when resizing
In isolated mode, resizing the frame is all that's requied to get the
body to change width, but in non-isolated mode, it doesn't change it's
width until the cycle yields. Setting the body to position relative
appears to force it to take the width immediately, resulting in a correct
value from getBodyHeight.

Change-Id: I8578ed2395dc91f166d72e6deb70130a1b42ebc1