ButtonGroupWidget: Remove weird margin-bottom: -1px; from theme styles
Added in 113e3ac1 a long time ago, causes all buttons in a button group
to be offset by one pixel to the normal position. Removing it doesn't
seem to affect the rendering of shadows and stuff. I don't know what
it was needed for, but doesn't seem needed anymore and breaks things.

Bug: T78041
Change-Id: If924415e96825157c2fcfae7e32938ef2ca06e58