FieldsetLayout: Shrink size of label and bump the weight to compensate
When we originally created the FieldsetLayout, the label was sized in
line with the rest of the themes. However, since then the drive has been
to have a more compact and clean look. Right now the FieldSetLayout
labels very significantly dominate the display of the dialogs which use
them.

Shrink the text size from 1.5em to 1.1em (linearly 27% smaller), bump
the weight to bold, and also shrink the padding around the icon a bit to
avoid skewing the baseline too much. There are no uses of the layouts in
the demo yet, sadly, so examining the change is non-trivial.

Bug: T78037
Change-Id: I7d69f43b5e6c33de82c10237d7d6db3c597b6bf6